+++ services/fetcher/mem/service.go
@@ -1,4 +1,4 @@
-// Copyright © 2020 Attestant Limited.
+// Copyright © 2020, 2021 Attestant Limited.
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -14,7 +14,6 @@
 package mem
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
@@ -27,6 +26,7 @@ import (
 	zerologger "github.com/rs/zerolog/log"
 	"github.com/wealdtech/go-bytesutil"
 	e2wallet "github.com/wealdtech/go-eth2-wallet"
+	distributed "github.com/wealdtech/go-eth2-wallet-distributed"
 	hd "github.com/wealdtech/go-eth2-wallet-hd/v2"
 	nd "github.com/wealdtech/go-eth2-wallet-nd/v2"
 	e2wtypes "github.com/wealdtech/go-eth2-wallet-types/v2"
@@ -34,15 +34,16 @@ import (
 
 // Service contains an in-memory cache of wallets and accounts.
 type Service struct {
-	monitor       metrics.FetcherMonitor
-	stores        []e2wtypes.Store
-	pubKeyPaths   map[[48]byte]string
-	pubKeyPathsMx sync.RWMutex
-	wallets       map[string]e2wtypes.Wallet
-	walletsMx     sync.RWMutex
-	accounts      map[string]e2wtypes.Account
-	accountsMx    sync.RWMutex
-	encryptor     e2wtypes.Encryptor
+	monitor        metrics.FetcherMonitor
+	pubKeyPaths    map[[48]byte]string
+	wallets        map[string]e2wtypes.Wallet
+	walletAccounts map[string]map[string]e2wtypes.Account
+	// Read-write copy of some information to allow for
+	// dynamic addition of accounts without requiring mutexes
+	// for normal access.
+	rwPubKeyPaths    map[[48]byte]string
+	rwWalletAccounts map[string]map[string]e2wtypes.Account
+	rwMu             sync.RWMutex
 }
 
 // module-wide log.
@@ -61,13 +62,18 @@ func New(ctx context.Context, params ...Parameter) (*Service, error) {
 		log = log.Level(parameters.logLevel)
 	}
 
+	wallets, walletAccounts, pubKeyPaths, err := populateCaches(ctx, parameters.stores, parameters.encryptor)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to populate caches")
+	}
+
 	s := &Service{
-		monitor:     parameters.monitor,
-		encryptor:   parameters.encryptor,
-		stores:      parameters.stores,
-		pubKeyPaths: make(map[[48]byte]string),
-		wallets:     make(map[string]e2wtypes.Wallet),
-		accounts:    make(map[string]e2wtypes.Account),
+		monitor:          parameters.monitor,
+		pubKeyPaths:      pubKeyPaths,
+		wallets:          wallets,
+		walletAccounts:   walletAccounts,
+		rwPubKeyPaths:    make(map[[48]byte]string),
+		rwWalletAccounts: make(map[string]map[string]e2wtypes.Account),
 	}
 
 	return s, nil
@@ -75,132 +81,171 @@ func New(ctx context.Context, params ...Parameter) (*Service, error) {
 
 // FetchWallet fetches the wallet.
 func (s *Service) FetchWallet(ctx context.Context, path string) (e2wtypes.Wallet, error) {
-	log := log.With().Str("path", path).Logger()
-	log.Trace().Msg("Fetching wallet")
-
 	walletName, _, err := e2wallet.WalletAndAccountNames(path)
 	if err != nil {
 		log.Warn().Msg("Invalid path")
 		return nil, errors.Wrap(err, "invalid path")
 	}
 
-	// Return wallet from cache if present.
-	s.walletsMx.RLock()
 	wallet, exists := s.wallets[walletName]
-	s.walletsMx.RUnlock()
 	if exists {
-		log.Trace().Msg("Wallet found in cache")
+		log.Trace().Str("wallet", walletName).Msg("Wallet found in cache")
 		return wallet, nil
 	}
+	log.Trace().Str("wallet", walletName).Msg("Wallet not found in cache")
 
-	for _, store := range s.stores {
-		wallet, err = e2wallet.OpenWallet(walletName, e2wallet.WithStore(store))
-		if err == nil {
-			break
-		}
-	}
-	if wallet == nil {
-		log.Warn().Msg("Wallet not found")
-		return nil, errors.New("wallet not found")
-	}
-
-	s.walletsMx.Lock()
-	s.wallets[walletName] = wallet
-	s.walletsMx.Unlock()
-
-	log.Trace().Msg("Wallet found in store")
-	return wallet, nil
+	return nil, errors.New("wallet not found")
 }
 
 // FetchAccount fetches the account given its name.
 func (s *Service) FetchAccount(ctx context.Context, path string) (e2wtypes.Wallet, e2wtypes.Account, error) {
-	log := log.With().Str("path", path).Logger()
-	log.Trace().Msg("Fetching account")
-
-	// Fetch account and store in cache if present.
-	wallet, err := s.FetchWallet(ctx, path)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	// Return account from cache if present.
-	s.accountsMx.RLock()
-	account, exists := s.accounts[path]
-	s.accountsMx.RUnlock()
-	if exists {
-		log.Trace().Msg("Account found in cache")
-		return wallet, account, nil
-	}
-
-	// Need to fetch manually
-	_, accountName, err := e2wallet.WalletAndAccountNames(path)
+	// Fetch the account name.
+	walletName, accountName, err := e2wallet.WalletAndAccountNames(path)
 	if err != nil {
 		log.Warn().Msg("Invalid path")
 		return nil, nil, errors.Wrap(err, "invalid path")
 	}
-	accountByNameProvider, isProvider := wallet.(e2wtypes.WalletAccountByNameProvider)
-	if !isProvider {
-		log.Warn().Msg("Account cannot be fetched by name")
-		return nil, nil, errors.New("wallet does not allow fetching account by name")
-	}
-	account, err = accountByNameProvider.AccountByName(ctx, accountName)
-	if err != nil {
-		log.Warn().Err(err).Msg("Account not found")
-		return nil, nil, errors.Wrap(err, "failed to obtain account by name")
-	}
-	s.accountsMx.Lock()
-	s.accounts[path] = account
-	s.accountsMx.Unlock()
-	s.pubKeyPathsMx.Lock()
-	s.pubKeyPaths[bytesutil.ToBytes48(account.(e2wtypes.AccountPublicKeyProvider).PublicKey().Marshal())] = fmt.Sprintf("%s/%s", wallet.Name(), account.Name())
-	s.pubKeyPathsMx.Unlock()
 
-	log.Trace().Msg("Account found in store")
-	return wallet, account, nil
+	wallet, exists := s.wallets[walletName]
+	if !exists {
+		return nil, nil, errors.New("failed to find wallet")
+	}
+
+	walletAccounts, exists := s.walletAccounts[walletName]
+	if exists {
+		account, exists := walletAccounts[accountName]
+		if exists {
+			return wallet, account, nil
+		}
+	}
+
+	// Try the rw cache.
+	s.rwMu.RLock()
+	defer s.rwMu.RUnlock()
+	walletAccounts, exists = s.rwWalletAccounts[walletName]
+	if exists {
+		account, exists := walletAccounts[accountName]
+		if exists {
+			return wallet, account, nil
+		}
+	}
+
+	return nil, nil, errors.New("failed to find account")
 }
 
 // FetchAccountByKey fetches the account given its public key.
 func (s *Service) FetchAccountByKey(ctx context.Context, pubKey []byte) (e2wtypes.Wallet, e2wtypes.Account, error) {
-	log := log.With().Str("public_key", fmt.Sprintf("%#x", pubKey)).Logger()
-	log.Trace().Msg("Fetching account by key")
-
-	// See if we already know this key.
-	s.pubKeyPathsMx.RLock()
-	account, exists := s.pubKeyPaths[bytesutil.ToBytes48(pubKey)]
-	s.pubKeyPathsMx.RUnlock()
-	if exists {
-		log.Trace().Msg("Account found in cache")
-		return s.FetchAccount(ctx, account)
-	}
-
-	// We don't.  Trawl wallets to find the result.
-	for _, store := range s.stores {
-		for walletBytes := range store.RetrieveWallets() {
-			wallet, err := walletFromBytes(ctx, walletBytes, store, s.encryptor)
-			if err != nil {
-				log.Error().Err(err).Msg("Failed to decode wallet")
-				continue
-			}
-			for account := range wallet.Accounts(ctx) {
-				if bytes.Equal(account.(e2wtypes.AccountPublicKeyProvider).PublicKey().Marshal(), pubKey) {
-					// Found it.
-					path := fmt.Sprintf("%s/%s", wallet.Name(), account.Name())
-					s.accountsMx.Lock()
-					s.accounts[path] = account
-					s.accountsMx.Unlock()
-					s.pubKeyPathsMx.Lock()
-					s.pubKeyPaths[bytesutil.ToBytes48(pubKey)] = path
-					s.pubKeyPathsMx.Unlock()
-
-					log.Trace().Msg("Account found in store")
-					return wallet, account, nil
-				}
-			}
+	path, exists := s.pubKeyPaths[bytesutil.ToBytes48(pubKey)]
+	if !exists {
+		s.rwMu.RLock()
+		path, exists = s.rwPubKeyPaths[bytesutil.ToBytes48(pubKey)]
+		s.rwMu.RUnlock()
+		if !exists {
+			return nil, nil, errors.New("public key not known")
 		}
 	}
 
-	log.Trace().Msg("Account not found")
-	return nil, nil, errors.New("account not found")
+	return s.FetchAccount(ctx, path)
+}
+
+// FetchAccounts fetches all accounts for the wallet.
+func (s *Service) FetchAccounts(ctx context.Context, path string) (map[string]e2wtypes.Account, error) {
+	// Fetch the wallet name.
+	walletName, _, err := e2wallet.WalletAndAccountNames(path)
+	if err != nil {
+		log.Warn().Msg("Invalid path")
+		return nil, errors.Wrap(err, "invalid path")
+	}
+
+	walletAccounts, exists := s.walletAccounts[walletName]
+	s.rwMu.RLock()
+	defer s.rwMu.RUnlock()
+	rwWalletAccounts, rwExists := s.rwWalletAccounts[walletName]
+	if !exists && !rwExists {
+		return nil, errors.New("found no accounts for wallet")
+	}
+	if rwExists {
+		allWalletAccounts := make(map[string]e2wtypes.Account, len(walletAccounts)+len(rwWalletAccounts))
+		for k, v := range walletAccounts {
+			allWalletAccounts[k] = v
+		}
+		for k, v := range rwWalletAccounts {
+			allWalletAccounts[k] = v
+		}
+		walletAccounts = allWalletAccounts
+	}
+
+	return walletAccounts, nil
+}
+
+// AddAccount adds an account to the fetcher's internal stores.
+func (s *Service) AddAccount(ctx context.Context, wallet e2wtypes.Wallet, account e2wtypes.Account) error {
+	s.rwMu.Lock()
+	defer s.rwMu.Unlock()
+	if _, exists := s.wallets[wallet.Name()]; !exists {
+		return errors.New("failed to find wallet")
+	}
+	if _, exists := s.rwWalletAccounts[wallet.Name()]; !exists {
+		s.rwWalletAccounts[wallet.Name()] = make(map[string]e2wtypes.Account)
+	}
+	s.rwWalletAccounts[wallet.Name()][account.Name()] = account
+
+	path := fmt.Sprintf("%s/%s", wallet.Name(), account.Name())
+	s.rwPubKeyPaths[bytesutil.ToBytes48(account.PublicKey().Marshal())] = path
+
+	return nil
+}
+
+// populateCaches populates wallet and account caches for the service.
+func populateCaches(ctx context.Context, stores []e2wtypes.Store, encryptor e2wtypes.Encryptor) (map[string]e2wtypes.Wallet, map[string]map[string]e2wtypes.Account, map[[48]byte]string, error) {
+	log.Trace().Msg("Populating fetcher caches")
+
+	wallets := make(map[string]e2wtypes.Wallet)
+	walletAccounts := make(map[string]map[string]e2wtypes.Account)
+	pubKeyPaths := make(map[[48]byte]string)
+	var mu sync.Mutex
+
+	var wg sync.WaitGroup
+	for _, store := range stores {
+		for walletBytes := range store.RetrieveWallets() {
+			wg.Add(1)
+			go func(
+				mu *sync.Mutex,
+				walletBytes []byte,
+				wallets map[string]e2wtypes.Wallet,
+				walletAccounts map[string]map[string]e2wtypes.Account,
+				pubKeyPaths map[[48]byte]string,
+				wg *sync.WaitGroup) {
+				defer wg.Done()
+				wallet, err := walletFromBytes(ctx, walletBytes, store, encryptor)
+				if err != nil {
+					log.Error().Err(err).Msg("failed to decode wallet")
+					return
+				}
+				mu.Lock()
+				wallets[wallet.Name()] = wallet
+				mu.Unlock()
+				log.Trace().Str("wallet", wallet.Name()).Msg("Found wallet")
+
+				// Add each individual accounts.
+				accounts := make(map[string]e2wtypes.Account)
+				for account := range wallet.Accounts(ctx) {
+					accounts[account.Name()] = account
+					path := fmt.Sprintf("%s/%s", wallet.Name(), account.Name())
+					mu.Lock()
+					pubKeyPaths[bytesutil.ToBytes48(account.PublicKey().Marshal())] = path
+					mu.Unlock()
+					log.Trace().Str("wallet", wallet.Name()).Str("account", account.Name()).Msg("Stored account")
+				}
+				mu.Lock()
+				walletAccounts[wallet.Name()] = accounts
+				mu.Unlock()
+			}(&mu, walletBytes, wallets, walletAccounts, pubKeyPaths, &wg)
+		}
+	}
+	wg.Wait()
+
+	return wallets, walletAccounts, pubKeyPaths, nil
 }
 
 func walletFromBytes(ctx context.Context, data []byte, store e2wtypes.Store, encryptor e2wtypes.Encryptor) (e2wtypes.Wallet, error) {
@@ -231,6 +276,8 @@ func walletFromBytes(ctx context.Context, data []byte, store e2wtypes.Store, enc
 		wallet, err = nd.DeserializeWallet(ctx, data, store, encryptor)
 	case "hd", "hierarchical deterministic":
 		wallet, err = hd.DeserializeWallet(ctx, data, store, encryptor)
+	case "distributed":
+		wallet, err = distributed.DeserializeWallet(ctx, data, store, encryptor)
 	default:
 		return nil, fmt.Errorf("unsupported wallet type %q", info.Type)
 	}