+++ testing/daemon/daemon.go
@@ -1,4 +1,4 @@
-// Copyright © 2020 Attestant Limited.
+// Copyright © 2020, 2021 Attestant Limited.
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -290,6 +290,7 @@ func New(ctx context.Context, path string, id uint64, port uint32, peersMap map[
 		standardprocess.WithChecker(checker),
 		standardprocess.WithUnlocker(unlocker),
 		standardprocess.WithSender(sender),
+		standardprocess.WithFetcher(fetcher),
 		standardprocess.WithEncryptor(keystorev4.New()),
 		standardprocess.WithPeers(peers),
 		standardprocess.WithID(id),