+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Development
   - garbage collect the slashing database on startup to reduce on-disk size
   - provide release metric in `dirk_release`
+  - use internal account cache for both positive and negative caching
 
 # Version 1.0.4
   - Update dependencies